Fix the exported functions for auth so next doesnt complain on newer versions
