+++ app/(pages)/active/page.tsx
@@ -1,6 +1,6 @@
 import { redirect } from "next/navigation";
 import { getServerSession } from "next-auth";
-import { authOptions } from "@/api/auth/[...nextauth]/route";
+import { authOptions } from "@/lib/authOptions";
 import ActiveItem from "./ActiveItem";
 
 interface item {
