+++ app/(pages)/create/page.tsx
@@ -1,6 +1,6 @@
 import CreateForm from "./CreateForm";
 import { getServerSession } from "next-auth";
-import { authOptions } from "@/api/auth/[...nextauth]/route";
+import { authOptions } from "@/lib/authOptions";
 import { redirect } from "next/navigation";
 
 const Create = async () => {
