+++ app/(pages)/edit/[id]/page.tsx
@@ -1,6 +1,6 @@
 import EditForm from "./EditForm";
 import { getServerSession } from "next-auth";
-import { authOptions } from "@/api/auth/[...nextauth]/route";
+import { authOptions } from "@/lib/authOptions";
 import { redirect } from "next/navigation";
 import { prisma } from "@/lib/prisma";
