+++ app/(pages)/moderation/moderate/[creatorId]/navbar.tsx
@@ -1,5 +1,5 @@
 import Link from "next/link";
-import { authOptions } from "@/api/auth/[...nextauth]/route";
+import { authOptions } from "@/lib/authOptions";
 import { prisma } from "@/lib/prisma";
 import { getServerSession } from "next-auth";
 import { redirect } from "next/navigation";
@@ -20,10 +20,7 @@ export default async function ModNavBar({ slug }: { slug: string }) {
     return (
         <>
             <div className="border flex flex-row justify-center m-2 p-2 rounded-lg">
-                <Link className="block p-2 cursor-pointer bg-slate-600 hover:bg-blue-700 rounded-xl text-center m-2" href={`/moderation/moderate/${slug}/banned/1`}>
-                    Banned
-                </Link>
-                <Link className="block p-2 cursor-pointer bg-slate-600 hover:bg-blue-700 rounded-xl text-center m-2" href={`/moderation/moderate/${slug}/viewers/1`}>
+                <Link className="block p-2 cursor-pointer bg-slate-600 hover:bg-blue-700 rounded-xl text-center m-2" href={`/moderation/moderate/${slug}/viewers`}>
                     Viewers
                 </Link>
                 <Link className="block p-2 cursor-pointer bg-slate-600 hover:bg-blue-700 rounded-xl text-center m-2" href={`/moderation/moderate/${slug}/winners/1`}>
