+++ app/(pages)/moderation/page.tsx
@@ -1,4 +1,4 @@
-import { authOptions } from "@/api/auth/[...nextauth]/route";
+import { authOptions } from "@/lib/authOptions";
 import { prisma } from "@/lib/prisma";
 import { getServerSession } from "next-auth";
 import Link from "next/link";
