+++ app/(pages)/wheel/[id]/page.tsx
@@ -1,6 +1,6 @@
 import RaffleUI from "./RaffleUI";
 import { getServerSession } from "next-auth";
-import { authOptions } from "@/api/auth/[...nextauth]/route";
+import { authOptions } from "@/lib/authOptions";
 import { redirect } from "next/navigation";
 import { prisma } from "@/lib/prisma";
