+++ app/api/raffle/route.ts
@@ -1,7 +1,7 @@
 import { prisma } from "@/lib/prisma";
 import { getServerSession } from "next-auth";
 import { NextRequest, NextResponse } from "next/server";
-import { authOptions } from "../auth/[...nextauth]/route";
+import { authOptions } from "@/lib/authOptions";
 import { createRewardsSub, deleteListener } from "../webhooks/_helpers";
 
 /** Create a new reward and save it to the db */
